Доработана сетевая часть для поиска вакансий. Реализован репозиторий и интерактор для поиска вакансий.

+++ app/src/main/java/ru/practicum/android/diploma/common/domain/model/Salary.kt
@@ -2,6 +2,6 @@ package ru.practicum.android.diploma.common.domain.model
 
 data class Salary(
     val currency: String?,
-    val from: Double?,
-    val to: Double?
+    val from: Int?,
+    val to: Int?
 )
